# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Parrot Translator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TranslateInput {
  /** @minLength 1 */
  text: string;
  source: string;
  target: string;
}

export interface TranslateResult {
  translatedText: string;
  originalText: string;
  source: string;
  target: string;
  /** @nullable */
  detectedLanguage?: string | null;
}

export interface DetectInput {
  /** @minLength 1 */
  text: string;
}

export interface DetectResult {
  language: string;
  confidence: number;
}

export interface HistoryItem {
  id: number;
  originalText: string;
  translatedText: string;
  sourceLanguage: string;
  targetLanguage: string;
  createdAt: string;
}

export interface HistoryInput {
  originalText: string;
  translatedText: string;
  sourceLanguage: string;
  targetLanguage: string;
}

export interface FavoriteItem {
  id: number;
  originalText: string;
  translatedText: string;
  sourceLanguage: string;
  targetLanguage: string;
  createdAt: string;
}

export interface FavoriteInput {
  originalText: string;
  translatedText: string;
  sourceLanguage: string;
  targetLanguage: string;
}

export interface PhraseOfDay {
  phrase: string;
  translation: string;
  sourceLanguage: string;
  targetLanguage: string;
  pronunciationHint: string;
  /** @nullable */
  funFact?: string | null;
}