# lib/api-zod/src/generated/types/detectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Parrot Translator API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectInput {
  /** @minLength 1 */
  text: string;
}