# lib/api-zod/src/generated/types/favoriteItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Parrot Translator API
 * OpenAPI spec version: 0.1.0
 */

export interface FavoriteItem {
  id: number;
  originalText: string;
  translatedText: string;
  sourceLanguage: string;
  targetLanguage: string;
  createdAt: string;
}