# lib/api-zod/src/generated/types/phraseOfDay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Parrot Translator API
 * OpenAPI spec version: 0.1.0
 */

export interface PhraseOfDay {
  phrase: string;
  translation: string;
  sourceLanguage: string;
  targetLanguage: string;
  pronunciationHint: string;
  /** @nullable */
  funFact?: string | null;
}