# lib/api-zod/src/generated/types/translateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Parrot Translator API
 * OpenAPI spec version: 0.1.0
 */

export interface TranslateResult {
  translatedText: string;
  originalText: string;
  source: string;
  target: string;
  /** @nullable */
  detectedLanguage?: string | null;
}